+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Remove dependency on download library (@mbjoseph #249)
 * Remove earthpy.utils.fix_paths as it is not used in the package (@lwasser #259)
 * Adding tests for hillshade and improved docs (@jpalomino #260)
 * Closing plots in tests (@lwasser #257)
-* Added a code of conduct (@mbjoseph, #27)
-* Added tests for EarthlabData class (@mbjoseph, #37)
+* Added a code of conduct (@mbjoseph #27)
+* Added tests for EarthlabData class (@mbjoseph #37)
 
 ## [0.6.2] - 2019-02-19
 We have made significant changes in preparation for a 1.0 release